# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Freightrek Server API",
    "version": "1.0.0",
    "description": "REST API for Freightrek freight management system with admin authentication and RBAC",
    "contact": {
      "name": "Freightrek Team",
      "email": "support@freightrek.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "https://freightrekapi.vercel.app",
      "description": "Production (Vercel)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development (local)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token obtained from login endpoint"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/admin/agency": {
      "post": {
        "summary": "Create a new agency",
        "tags": [
          "Agency Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Agency"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agency created successfully"
          },
          "400": {
            "description": "Validation error or agency already exists"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Permission denied"
          }
        }
      },
      "get": {
        "summary": "Get all agencies with pagination and filters",
        "tags": [
          "Agency Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "Active",
                "Inactive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of agencies"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Permission denied"
          }
        }
      }
    },
    "/admin/agency/{id}": {
      "get": {
        "summary": "Get agency by ID",
        "tags": [
          "Agency Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agency details"
          },
          "404": {
            "description": "Agency not found"
          }
        }
      },
      "put": {
        "summary": "Update agency",
        "tags": [
          "Agency Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Agency"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agency updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Agency not found"
          }
        }
      },
      "delete": {
        "summary": "Delete agency (staff are auto-reassigned to another agency)",
        "description": "Any agency staff assigned to this agency are moved to another active agency before the agency is removed. The target agency is picked automatically (same city, then same state, then any active agency) unless reassignAgencyId is supplied. Franchise-scoped roles are remapped to the same role name in the target agency, or cleared if it has none. The delete is rejected when staff exist and no other active agency is available.\n",
        "tags": [
          "Agency Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "reassignAgencyId",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Agency to move this agency's staff to. Auto-selected when omitted."
          }
        ],
        "responses": {
          "200": {
            "description": "Agency deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "reassignedStaffCount": {
                          "type": "number"
                        },
                        "reassignedToAgency": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No active agency available for the staff, or invalid reassignAgencyId"
          },
          "404": {
            "description": "Agency not found"
          }
        }
      }
    },
...
    "/admin/agency/invoice": {
      "get": {
        "summary": "This agency's invoices",
        "tags": [
          "Invoice - Agency"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "issued",
                "cancelled"
              ]
            }
          },
          {
            "in": "query",
            "name": "orderId",
            "schema": {
              "type": "string"
            },
            "description": "Parcel order ObjectId (alias: order)"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dateFrom",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "dateTo",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoices plus totals for this agency"
          },
          "403": {
            "description": "Agency access required"
          }
        }
      }
    },
    "/admin/agency/invoice/summary": {
      "get": {
        "summary": "This agency's billing summary",
        "tags": [
          "Invoice - Agency"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "dateFrom",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "dateTo",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Totals billed by this agency"
          }
        }
      }
    },
    "/admin/agency/invoice/number/{invoiceNumber}": {
      "get": {
        "summary": "Look up one of this agency's invoices by number",
        "tags": [
          "Invoice - Agency"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "invoiceNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice details"
          },
          "403": {
            "description": "Invoice belongs to another agency"
          }
        }
      }
    },
    "/admin/agency/invoice/order/{orderId}": {
      "get": {
        "summary": "The invoice for one of this agency's orders",
        "tags": [
          "Invoice - Agency"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice details"
          },
          "403": {
            "description": "Invoice belongs to another agency"
          }
        }
      }
    },
    "/admin/agency/invoice/order/{orderId}/generate": {
      "post": {
        "summary": "Raise an invoice for one of this agency's orders",
        "description": "New bookings are invoiced automatically. Use this for orders booked before invoicing existed, or to re-issue after a cancellation (a fresh invoice number is allocated). Only orders booked at this agency are accepted.\n",
        "tags": [
          "Invoice - Agency"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Invoice raised"
          },
          "403": {
            "description": "Order was booked at another agency"
          },
          "404": {
            "description": "Parcel order not found"
          },
          "409": {
            "description": "An issued invoice already exists"
          }
        }
      }
    },
    "/admin/agency/invoice/{id}": {
      "get": {
        "summary": "Get one of this agency's invoices",
        "tags": [
          "Invoice - Agency"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice details"
          },
          "403": {
            "description": "Invoice belongs to another agency"
          }
        }
      }
    },
...
    "/admin/invoice": {
      "get": {
        "summary": "List invoices",
        "tags": [
          "Invoice"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "agency",
            "schema": {
              "type": "string"
            },
            "description": "Agency ObjectId"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "issued",
                "cancelled"
              ]
            }
          },
          {
            "in": "query",
            "name": "invoiceNumber",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "orderId",
            "schema": {
              "type": "string"
            },
            "description": "Parcel order ObjectId (alias: order). Returns the single invoice of that order"
          },
          {
            "in": "query",
            "name": "orderNumber",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "paymentType",
            "schema": {
              "type": "string",
              "enum": [
                "Paid",
                "To Pay",
                "Credit"
              ]
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Matches invoice/order number, customer name, mobile or GST"
          },
          {
            "in": "query",
            "name": "dateFrom",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "dateTo",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoices plus totals — totalTransportation, totalLoading, totalMiscellaneous, totalAmount\n"
          }
        }
      }
    },
    "/admin/invoice/summary": {
      "get": {
        "summary": "Billing summary, overall and per agency",
        "tags": [
          "Invoice"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "agency",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dateFrom",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "dateTo",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "issuedInvoices, totalTransportation, totalLoading, totalMiscellaneous, totalAmount, cancelledInvoices, perAgency[]\n"
          }
        }
      }
    },
    "/admin/invoice/number/{invoiceNumber}": {
      "get": {
        "summary": "Look an invoice up by its number",
        "tags": [
          "Invoice"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "invoiceNumber",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "INV-2026-27-000001"
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice details"
          },
          "404": {
            "description": "Invoice not found"
          }
        }
      }
    },
    "/admin/invoice/order/{orderId}": {
      "get": {
        "summary": "The invoice raised for a parcel order",
        "tags": [
          "Invoice"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice details"
          },
          "404": {
            "description": "No invoice for this order"
          }
        }
      }
    },
    "/admin/invoice/order/{orderId}/generate": {
      "post": {
        "summary": "Raise an invoice for an order that has none",
        "description": "New bookings are invoiced automatically. Use this for orders booked before invoicing existed, or to re-issue after a cancellation (a fresh invoice number is allocated).\n",
        "tags": [
          "Invoice"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Invoice raised"
          },
          "404": {
            "description": "Parcel order not found"
          },
          "409": {
            "description": "An issued invoice already exists"
          }
        }
      }
    },
    "/admin/invoice/{id}": {
      "get": {
        "summary": "Get an invoice by ID",
        "tags": [
          "Invoice"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice details"
          },
          "404": {
            "description": "Invoice not found"
          }
        }
      },
      "patch": {
        "summary": "Edit invoice notes",
        "description": "Only the notes are editable. Amounts follow the order's charges — change those with PATCH /admin/parcel-order/{id}/charge and the invoice is revised automatically, keeping a revision trail.\n",
        "tags": [
          "Invoice"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "notes"
                ],
                "properties": {
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notes updated"
          }
        }
      }
    },
...
    "/admin/route": {
      "post": {
        "summary": "Create a new route",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "routeName",
                  "from",
                  "to"
                ],
                "properties": {
                  "routeName": {
                    "type": "string",
                    "example": "Chennai - Bangalore Express"
                  },
                  "from": {
                    "type": "string",
                    "example": "Chennai"
                  },
                  "to": {
                    "type": "string",
                    "example": "Bangalore"
                  },
                  "branches": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "Guindy",
                      "Tambaram"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "Active",
                      "Inactive"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Route created successfully"
          },
          "400": {
            "description": "Validation error or route already exists"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Permission denied"
          }
        }
      },
      "get": {
        "summary": "Get all routes with pagination and filters",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "Active",
                "Inactive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of routes"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Permission denied"
          }
        }
      }
    },
    "/admin/route/{id}": {
      "get": {
        "summary": "Get route by ID",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route details"
          },
          "404": {
            "description": "Route not found"
          }
        }
      },
      "put": {
        "summary": "Update route",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "routeName": {
                    "type": "string"
                  },
                  "from": {
                    "type": "string"
                  },
                  "to": {
                    "type": "string"
                  },
                  "branches": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "Active",
                      "Inactive"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Route updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Route not found"
          }
        }
      },
      "delete": {
        "summary": "Delete route",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route deleted successfully"
          },
          "404": {
            "description": "Route not found"
          }
        }
      }
    },
...
    "/hub/invoice/summary": {
      "get": {
        "summary": "Billing summary for the parcels routed through this hub",
        "tags": [
          "Invoice - Hub"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "dateFrom",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "dateTo",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Totals",
            "overall and per agency": null
          }
        }
      }
    },
    "/hub/invoice/number/{invoiceNumber}": {
      "get": {
        "summary": "Get one invoice by its number",
        "tags": [
          "Invoice - Hub"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "invoiceNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice"
          },
          "403": {
            "description": "The parcel is not assigned to this hub"
          },
          "404": {
            "description": "Invoice not found"
          }
        }
      }
    },
    "/hub/invoice/order/{orderId}": {
      "get": {
        "summary": "Get the invoice raised for one parcel order",
        "tags": [
          "Invoice - Hub"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice"
          },
          "403": {
            "description": "The parcel is not assigned to this hub"
          },
          "404": {
            "description": "Invoice not found"
          }
        }
      }
    },
...
  },
...
}